requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
opencv-python>=4.8.0
albumentations>=1.3.1
tqdm>=4.66.0
pyyaml>=6.0
